+++ QBCore_2F0666.base/resources/[qb]/qb-garages/config.lua
@@ -17,182 +17,182 @@ Config.VehicleClass = {
 
 Config.Garages = {
     motelgarage = {
-        label = 'Motel Parking',
+        label = 'Motel Otoparkı',
         takeVehicle = vector3(274.29, -334.15, 44.92),
         spawnPoint = {
             vector4(265.96, -332.3, 44.51, 250.68)
         },
         showBlip = true,
-        blipName = 'Public Parking',
+        blipName = 'Açık Otopark',
         blipNumber = 357,
         blipColor = 3,
         type = 'public', -- public, gang, job, depot
         category = Config.VehicleClass['car']
     },
     casinogarage = {
-        label = 'Casino Parking',
+        label = 'Kumarhane Otoparkı',
         takeVehicle = vector3(883.96, -4.71, 78.76),
         spawnPoint = {
             vector4(895.39, -4.75, 78.35, 146.85)
         },
         showBlip = true,
-        blipName = 'Public Parking',
+        blipName = 'Açık Otopark',
         blipNumber = 357,
         blipColor = 3,
         type = 'public',
         category = Config.VehicleClass['car']
     },
     sapcounsel = {
-        label = 'San Andreas Parking',
+        label = 'San Andreas Otoparkı',
         takeVehicle = vector3(-330.01, -780.33, 33.96),
         spawnPoint = {
             vector4(-341.57, -767.45, 33.56, 92.61)
         },
         showBlip = true,
-        blipName = 'Public Parking',
+        blipName = 'Açık Otopark',
         blipNumber = 357,
         blipColor = 3,
         type = 'public',
         category = Config.VehicleClass['car']
     },
     spanishave = {
-        label = 'Spanish Ave Parking',
+        label = 'Spanish Ave Otoparkı',
         takeVehicle = vector3(-1160.86, -741.41, 19.63),
         spawnPoint = {
             vector4(-1145.2, -745.42, 19.26, 108.22)
         },
         showBlip = true,
-        blipName = 'Public Parking',
+        blipName = 'Açık Otopark',
         blipNumber = 357,
         blipColor = 3,
         type = 'public',
         category = Config.VehicleClass['car']
     },
     caears24 = {
-        label = 'Caears 24 Parking',
+        label = 'Caears 24 Otoparkı',
         takeVehicle = vector3(69.84, 12.6, 68.96),
         spawnPoint = {
             vector4(60.8, 17.54, 68.82, 339.7)
         },
         showBlip = true,
-        blipName = 'Public Parking',
+        blipName = 'Açık Otopark',
         blipNumber = 357,
         blipColor = 3,
         type = 'public',
         category = Config.VehicleClass['car']
     },
     caears242 = {
-        label = 'Caears 24 Parking',
+        label = 'Caears 24 Otoparkı',
         takeVehicle = vector3(-453.7, -786.78, 30.56),
         spawnPoint = {
             vector4(-472.39, -787.71, 30.14, 180.52)
         },
         showBlip = true,
-        blipName = 'Public Parking',
+        blipName = 'Açık Otopark',
         blipNumber = 357,
         blipColor = 3,
         type = 'public',
         category = Config.VehicleClass['car']
     },
     lagunapi = {
-        label = 'Laguna Parking',
+        label = 'Laguna Otoparkı',
         takeVehicle = vector3(364.37, 297.83, 103.49),
         spawnPoint = {
             vector4(375.09, 294.66, 102.86, 164.04)
         },
         showBlip = true,
-        blipName = 'Public Parking',
+        blipName = 'Açık Otopark',
         blipNumber = 357,
         blipColor = 3,
         type = 'public',
         category = Config.VehicleClass['car']
     },
     airportp = {
-        label = 'Airport Parking',
+        label = 'Havalimanı Otoparkı',
         takeVehicle = vector3(-773.12, -2033.04, 8.88),
         spawnPoint = {
             vector4(-779.77, -2040.18, 8.47, 315.34)
         },
         showBlip = true,
-        blipName = 'Public Parking',
+        blipName = 'Açık Otopark',
         blipNumber = 357,
         blipColor = 3,
         type = 'public',
         category = Config.VehicleClass['car']
     },
     beachp = {
-        label = 'Beach Parking',
+        label = 'Sahil Otoparkı',
         takeVehicle = vector3(-1185.32, -1500.64, 4.38),
         spawnPoint = {
             vector4(-1188.14, -1487.95, 3.97, 124.06)
         },
         showBlip = true,
-        blipName = 'Public Parking',
+        blipName = 'Açık Otopark',
         blipNumber = 357,
         blipColor = 3,
         type = 'public',
         category = Config.VehicleClass['car']
     },
     themotorhotel = {
-        label = 'The Motor Hotel Parking',
+        label = 'Motor Hotel Otoparkı',
         takeVehicle = vector3(1137.77, 2663.54, 37.9),
         spawnPoint = {
             vector4(1127.7, 2647.84, 37.58, 1.41)
         },
         showBlip = true,
-        blipName = 'Public Parking',
+        blipName = 'Açık Otopark',
         blipNumber = 357,
         blipColor = 3,
         type = 'public',
         category = Config.VehicleClass['car']
     },
     liqourparking = {
-        label = 'Liqour Parking',
+        label = 'Market Otoparkı',
         takeVehicle = vector3(883.99, 3649.67, 32.87),
         spawnPoint = {
             vector4(898.38, 3649.41, 32.36, 90.75)
         },
         showBlip = true,
-        blipName = 'Public Parking',
+        blipName = 'Açık Otopark',
         blipNumber = 357,
         blipColor = 3,
         type = 'public',
         category = Config.VehicleClass['car']
     },
     shoreparking = {
-        label = 'Shore Parking',
+        label = 'Kıyı Otoparkı',
         takeVehicle = vector3(1737.03, 3718.88, 34.05),
         spawnPoint = {
             vector4(1725.4, 3716.78, 34.15, 20.54)
         },
         showBlip = true,
-        blipName = 'Public Parking',
+        blipName = 'Açık Otopark',
         blipNumber = 357,
         blipColor = 3,
         type = 'public',
         category = Config.VehicleClass['car']
     },
     haanparking = {
-        label = 'Bell Farms Parking',
+        label = 'Bell Farms Otoparkı',
         takeVehicle = vector3(76.88, 6397.3, 31.23),
         spawnPoint = {
             vector4(62.15, 6403.41, 30.81, 211.38)
         },
         showBlip = true,
-        blipName = 'Public Parking',
+        blipName = 'Açık Otopark',
         blipNumber = 357,
         blipColor = 3,
         type = 'public',
         category = Config.VehicleClass['car']
     },
     dumbogarage = {
-        label = 'Dumbo Private Parking',
+        label = 'Dumbo Özel Otopark',
         takeVehicle = vector3(165.75, -3227.2, 5.89),
         spawnPoint = {
             vector4(168.34, -3236.1, 5.43, 272.05)
         },
         showBlip = true,
-        blipName = 'Public Parking',
+        blipName = 'Açık Otopark',
         blipNumber = 357,
         blipColor = 3,
         type = 'public',
@@ -209,33 +209,33 @@ Config.Garages = {
             vector4(234.42, -802.76, 30.04, 67.2)
         },
         showBlip = true,
-        blipName = 'Public Parking',
+        blipName = 'Açık Otopark',
         blipNumber = 357,
         blipColor = 3,
         type = 'public',
         category = Config.VehicleClass['car']
     },
     grapeseedgarage = {
-        label = 'Grapeseed Parking',
+        label = 'Grapeseed Otoparkı',
         takeVehicle = vector3(2552.68, 4671.8, 33.95),
         spawnPoint = {
             vector4(2550.17, 4681.96, 33.81, 17.05)
         },
         showBlip = true,
-        blipName = 'Public Parking',
+        blipName = 'Açık Otopark',
         blipNumber = 357,
         blipColor = 3,
         type = 'public',
         category = Config.VehicleClass['car']
     },
     depotLot = {
-        label = 'Depot Lot',
+        label = 'Çekilmiş Araçlar',
         takeVehicle = vector3(401.76, -1632.57, 29.29),
         spawnPoint = {
             vector4(396.55, -1643.93, 28.88, 321.91)
         },
         showBlip = true,
-        blipName = 'Depot Lot',
+        blipName = 'Çekilmiş Araçlar',
         blipNumber = 68,
         blipColor = 3,
         type = 'depot',
@@ -302,13 +302,13 @@ Config.Garages = {
         jobType = 'cartel'
     },
     police = {
-        label = 'Police',
+        label = 'Polis Garajı',
         takeVehicle = vector3(462.83, -1019.52, 28.1),
         spawnPoint = {
             vector4(446.16, -1025.79, 28.23, 6.59)
         },
         showBlip = false,
-        blipName = 'Police',
+        blipName = 'Polis Garajı',
         blipNumber = 357,
         blipColor = 3,
         type = 'job',
@@ -317,7 +317,7 @@ Config.Garages = {
         jobType = 'leo'
     },
     intairport = {
-        label = 'Airport Hangar',
+        label = 'Havalimanı Hangarı',
         takeVehicle = vector3(-979.06, -2995.48, 13.95),
         spawnPoint = {
             vector4(-998.37, -2985.01, 13.95, 61.09)
@@ -344,7 +344,7 @@ Config.Garages = {
         category = Config.VehicleClass['air']
     },
     airsshores = {
-        label = 'Sandy Shores Hangar',
+        label = 'Sandy Shores Hangarı',
         takeVehicle = vector3(1737.89, 3288.13, 41.14),
         spawnPoint = {
             vector4(1742.83, 3266.83, 41.24, 102.64)
@@ -357,7 +357,7 @@ Config.Garages = {
         category = Config.VehicleClass['air']
     },
     airzancudo = {
-        label = 'Fort Zancudo Hangar',
+        label = 'Fort Zancudo Hangarı',
         takeVehicle = vector3(-1828.25, 2975.44, 32.81),
         spawnPoint = {
             vector4(-1828.25, 2975.44, 32.81, 57.24)
@@ -370,143 +370,143 @@ Config.Garages = {
         category = Config.VehicleClass['air']
     },
     airdepot = {
-        label = 'Air Depot',
+        label = 'Hava Aracı Deposu',
         takeVehicle = vector3(-1270.01, -3377.53, 14.33),
         spawnPoint = {
             vector4(-1270.01, -3377.53, 14.33, 329.25)
         },
         showBlip = true,
-        blipName = 'Air Depot',
+        blipName = 'Hava Deposu',
         blipNumber = 359,
         blipColor = 3,
         type = 'depot',
         category = Config.VehicleClass['air']
     },
     lsymc = {
-        label = 'LSYMC Boathouse',
+        label = 'LSYMC Tekne Garajı',
         takeVehicle = vector3(-785.95, -1497.84, -0.09),
         spawnPoint = {
             vector4(-796.64, -1502.6, -0.09, 111.49)
         },
         showBlip = true,
-        blipName = 'Boathouse',
+        blipName = 'Tekne Garajı',
         blipNumber = 356,
         blipColor = 3,
         type = 'public',
         category = Config.VehicleClass['sea']
     },
     paleto = {
-        label = 'Paleto Boathouse',
+        label = 'Paleto Tekne Garajı',
         takeVehicle = vector3(-278.21, 6638.13, 7.55),
         spawnPoint = {
             vector4(-289.2, 6637.96, 1.01, 45.5)
         },
         showBlip = true,
-        blipName = 'Boathouse',
+        blipName = 'Tekne Garajı',
         blipNumber = 356,
         blipColor = 3,
         type = 'public',
         category = Config.VehicleClass['sea']
     },
     millars = {
-        label = 'Millars Boathouse',
+        label = 'Millars Tekne Garajı',
         takeVehicle = vector3(1298.56, 4212.42, 33.25),
         spawnPoint = {
             vector4(1297.82, 4209.61, 30.12, 253.5)
         },
         showBlip = true,
-        blipName = 'Boathouse',
+        blipName = 'Tekne Garajı',
         blipNumber = 356,
         blipColor = 3,
         type = 'public',
         category = Config.VehicleClass['sea']
     },
     seadepot = {
-        label = 'LSYMC Depot',
+        label = 'LSYMC Tekne Deposu',
         takeVehicle = vector3(-742.95, -1407.58, 5.5),
         spawnPoint = {
             vector4(-729.77, -1355.49, 1.19, 142.5)
         },
         showBlip = true,
-        blipName = 'LSYMC Depot',
+        blipName = 'Tekne Deposu',
         blipNumber = 356,
         blipColor = 3,
         type = 'depot',
         category = Config.VehicleClass['sea']
     },
     rigdepot = {
-        label = 'Big Rig Depot',
+        label = 'TIR Deposu',
         takeVehicle = vector3(2334.42, 3118.62, 48.2),
         spawnPoint = {
             vector4(2324.57, 3117.79, 48.21, 4.05)
         },
         showBlip = true,
-        blipName = 'Big Rig Depot',
+        blipName = 'TIR Deposu',
         blipNumber = 68,
         blipColor = 2,
         type = 'depot',
         category = Config.VehicleClass['rig']
     },
     dumborigparking = {
-        label = 'Dumbo Big Rig Parking',
+        label = 'Dumbo TIR Otoparkı',
         takeVehicle = vector3(161.23, -3188.73, 5.97),
         spawnPoint = {
             vector4(167.0, -3203.89, 5.94, 271.27)
         },
         showBlip = true,
-        blipName = 'Big Rig Parking',
+        blipName = 'TIR Otoparkı',
         blipNumber = 357,
         blipColor = 2,
         type = 'public',
         category = Config.VehicleClass['rig']
     },
     popsrigparking = {
-        label = 'Pop\'s Big Rig Parking',
+        label = 'Pop\'s TIR Otoparkı',
         takeVehicle = vector3(137.67, 6632.99, 31.67),
         spawnPoint = {
             vector4(127.69, 6605.84, 31.93, 223.67)
         },
         showBlip = true,
-        blipName = 'Big Rig Parking',
+        blipName = 'TIR Otoparkı',
         blipNumber = 357,
         blipColor = 2,
         type = 'public',
         category = Config.VehicleClass['rig']
     },
     ronsrigparking = {
-        label = 'Ron\'s Big Rig Parking',
+        label = 'Ron\'s TIR Otoparkı',
         takeVehicle = vector3(-2529.37, 2342.67, 33.06),
         spawnPoint = {
             vector4(-2521.61, 2326.45, 33.13, 88.7)
         },
         showBlip = true,
-        blipName = 'Big Rig Parking',
+        blipName = 'TIR Otoparkı',
         blipNumber = 357,
         blipColor = 2,
         type = 'public',
         category = Config.VehicleClass['rig']
     },
     ronsrigparking2 = {
-        label = 'Ron\'s Big Rig Parking',
+        label = 'Ron\'s TIR Otoparkı',
         takeVehicle = vector3(2561.67, 476.68, 108.49),
         spawnPoint = {
             vector4(2561.67, 476.68, 108.49, 177.86)
         },
         showBlip = true,
-        blipName = 'Big Rig Parking',
+        blipName = 'TIR Otoparkı',
         blipNumber = 357,
         blipColor = 2,
         type = 'public',
         category = Config.VehicleClass['rig']
     },
     ronsrigparking3 = {
-        label = 'Ron\'s Big Rig Parking',
+        label = 'Ron\'s TIR Otoparkı',
         takeVehicle = vector3(-41.24, -2550.63, 6.01),
         spawnPoint = {
             vector4(-39.39, -2527.81, 6.08, 326.18)
         },
         showBlip = true,
-        blipName = 'Big Rig Parking',
+        blipName = 'TIR Otoparkı',
         blipNumber = 357,
         blipColor = 2,
         type = 'public',